+++ src/vehicle.cpp
@@ -1,4 +1,3 @@
-
 #include <algorithm>
 #include <iostream>
 #include "vehicle.h"
@@ -9,7 +8,7 @@
 
 Vehicle::Vehicle() = default;
 
-Vehicle::Vehicle(float s, float d, float v, float a, float yaw, string state) {
+Vehicle::Vehicle(double s, double d, double v, double a, double yaw, string state) {
 
   this->s = s;
   this->d = d;
@@ -22,7 +21,7 @@ Vehicle::Vehicle(float s, float d, float v, float a, float yaw, string state) {
 }
 
 vector<Vehicle> Vehicle::choose_next_state(map<int, vector<Vehicle>> predictions,
-                                           std::function<float(Vehicle, map<int, vector<Vehicle>>, vector<Vehicle>)> calculate_cost) {
+                                           std::function<double(Vehicle, map<int, vector<Vehicle>>, vector<Vehicle>)> calculate_cost) {
   /*
   Here you can implement the transition_function code from the Behavior Planning Pseudocode
   classroom concept. Your goal will be to return the best (lowest cost) trajectory corresponding
@@ -35,8 +34,8 @@ vector<Vehicle> Vehicle::choose_next_state(map<int, vector<Vehicle>> predictions
 
   */
   vector<string> states = successor_states();
-  float cost;
-  vector<float> costs;
+  double cost;
+  vector<double> costs;
   vector<string> final_states;
   vector<vector<Vehicle>> final_trajectories;
 
@@ -98,16 +97,16 @@ vector<Vehicle> Vehicle::generate_trajectory(string state, map<int, vector<Vehic
   return trajectory;
 }
 
-vector<float> Vehicle::get_kinematics(map<int, vector<Vehicle>> predictions, int lane) {
+vector<double> Vehicle::get_kinematics(map<int, vector<Vehicle>> predictions, int lane) {
   /*
   Gets next timestep kinematics (position, velocity, acceleration)
   for a given lane. Tries to choose the maximum velocity and acceleration,
   given other vehicle positions and accel/velocity constraints.
   */
-  float max_velocity_accel_limit = this->max_acceleration + this->v;
-  float new_position;
-  float new_velocity;
-  float new_accel;
+  double max_velocity_accel_limit = this->max_acceleration + this->v;
+  double new_position;
+  double new_velocity;
+  double new_accel;
   Vehicle vehicle_ahead;
   Vehicle vehicle_behind;
 
@@ -116,7 +115,7 @@ vector<float> Vehicle::get_kinematics(map<int, vector<Vehicle>> predictions, int
     if (get_vehicle_behind(predictions, lane, vehicle_behind)) {
       new_velocity = vehicle_ahead.v; //must travel at the speed of traffic, regardless of preferred buffer
     } else {
-      auto max_velocity_in_front = (float) ((vehicle_ahead.s - this->s - this->preferred_buffer) + vehicle_ahead.v - 0.5 * (this->a));
+      auto max_velocity_in_front = (double) ((vehicle_ahead.s - this->s - this->preferred_buffer) + vehicle_ahead.v - 0.5 * (this->a));
       new_velocity = min(min(max_velocity_in_front, max_velocity_accel_limit), this->target_speed);
     }
   } else {
@@ -124,7 +123,7 @@ vector<float> Vehicle::get_kinematics(map<int, vector<Vehicle>> predictions, int
   }
 
   new_accel = new_velocity - this->v; //Equation: (v_1 - v_0)/t = acceleration
-  new_position = this->s + new_velocity + new_accel / (float) 2.0;
+  new_position = this->s + new_velocity + new_accel / (double) 2.0;
   return{new_position, new_velocity, new_accel};
 
 }
@@ -133,7 +132,7 @@ vector<Vehicle> Vehicle::constant_speed_trajectory() {
   /*
   Generate a constant speed trajectory.
   */
-  float next_pos = position_at(1);
+  double next_pos = position_at(1);
   vector<Vehicle> trajectory = {Vehicle(this->s, this->d, this->v, this->a, 0, this->state),
                                 Vehicle(next_pos, this->d, this->v, 0, 0, this->state)};
   return trajectory;
@@ -144,10 +143,10 @@ vector<Vehicle> Vehicle::keep_lane_trajectory(map<int, vector<Vehicle>> predicti
   Generate a keep lane trajectory.
   */
   vector<Vehicle> trajectory = {Vehicle(this->s, this->d, this->v, this->a, 0, state)};
-  vector<float> kinematics = get_kinematics(predictions, (int) round(this->get_lane()) );
-  float new_s = kinematics[0];
-  float new_v = kinematics[1];
-  float new_a = kinematics[2];
+  vector<double> kinematics = get_kinematics(predictions, (int) round(this->get_lane()) );
+  double new_s = kinematics[0];
+  double new_v = kinematics[1];
+  double new_a = kinematics[2];
   trajectory.emplace_back(new_s, this->d, new_v, new_a, 0, "KL");
   return trajectory;
 }
@@ -156,13 +155,13 @@ vector<Vehicle> Vehicle::prep_lane_change_trajectory(string state, map<int, vect
   /*
   Generate a trajectory preparing for a lane change.
   */
-  float new_s;
-  float new_v;
-  float new_a;
+  double new_s;
+  double new_v;
+  double new_a;
   Vehicle vehicle_behind;
   int new_lane = this->get_lane() + lane_direction[state];
   vector<Vehicle> trajectory = {Vehicle(this->s, this->d, this->v, this->a, 0, this->state)};
-  vector<float> curr_lane_new_kinematics = get_kinematics(predictions, this->get_lane());
+  vector<double> curr_lane_new_kinematics = get_kinematics(predictions, this->get_lane());
 
   if (get_vehicle_behind(predictions, this->get_lane(), vehicle_behind)) {
     //Keep speed of current lane so as not to collide with car behind.
@@ -171,8 +170,8 @@ vector<Vehicle> Vehicle::prep_lane_change_trajectory(string state, map<int, vect
     new_a = curr_lane_new_kinematics[2];
 
   } else {
-    vector<float> best_kinematics;
-    vector<float> next_lane_new_kinematics = get_kinematics(predictions, new_lane);
+    vector<double> best_kinematics;
+    vector<double> next_lane_new_kinematics = get_kinematics(predictions, new_lane);
     //Choose kinematics with lowest velocity.
     if (next_lane_new_kinematics[1] < curr_lane_new_kinematics[1]) {
       best_kinematics = next_lane_new_kinematics;
@@ -204,7 +203,7 @@ vector<Vehicle> Vehicle::lane_change_trajectory(string state, map<int, vector<Ve
     }
   }
   trajectory.emplace_back(this->s, this->d, this->v, this->a, 0, this->state);
-  vector<float> kinematics = get_kinematics(predictions, new_lane);
+  vector<double> kinematics = get_kinematics(predictions, new_lane);
   trajectory.emplace_back(kinematics[0], this->d, kinematics[1], kinematics[2], 0, state);
   return trajectory;
 }
@@ -213,8 +212,8 @@ void Vehicle::increment(int dt = 1) {
   this->s = position_at(dt);
 }
 
-float Vehicle::position_at(int t) {
-  return this->s + this->v*t + this->a*t*t/(float)2.0;
+double Vehicle::position_at(int t) {
+  return this->s + this->v*t + this->a*t*t/(double)2.0;
 }
 
 bool Vehicle::get_vehicle_behind(map<int, vector<Vehicle>> other_cars, int lane, Vehicle & rVehicle) {
@@ -222,7 +221,7 @@ bool Vehicle::get_vehicle_behind(map<int, vector<Vehicle>> other_cars, int lane,
   Returns a true if a vehicle is found behind the current vehicle, false otherwise. The passed reference
   rVehicle is updated if a vehicle is found.
   */
-  float max_s = -1;
+  double max_s = -1;
   bool found_vehicle = false;
   Vehicle temp_vehicle;
   for (auto &it : other_cars) {
@@ -243,7 +242,7 @@ bool Vehicle::get_vehicle_ahead(map<int, vector<Vehicle>> predictions, int lane,
   Returns a true if a vehicle is found ahead of the current vehicle, false otherwise. The passed reference
   rVehicle is updated if a vehicle is found.
   */
-  float min_s = this->target_s;
+  double min_s = this->target_s;
   bool found_vehicle = false;
   Vehicle temp_vehicle;
   for (auto &prediction : predictions) {
@@ -265,8 +264,8 @@ vector<Vehicle> Vehicle::generate_predictions(int horizon) {
   */
   vector<Vehicle> predictions;
   for(int i = 0; i < horizon; i++) {
-    float next_s = position_at(i);
-    float next_v = 0;
+    double next_s = position_at(i);
+    double next_v = 0;
     if (i < horizon-1) {
       next_v = position_at(i+1) - s;
     }
@@ -288,17 +287,17 @@ void Vehicle::realize_next_state(vector<Vehicle> trajectory) {
   //this->a = next_state.a;
 }
 
-void Vehicle::configure(vector<float> road_data) {
+void Vehicle::configure(vector<double> road_data) {
   /*
   Called by simulator before simulation begins. Sets various
   parameters which will impact the ego vehicle.
   */
-  target_speed = (float) road_data[0];
+  target_speed = (double) road_data[0];
   lanes_available = (int) road_data[1];
-  target_s = (float) road_data[2];
+  target_s = (double) road_data[2];
   target_lane = (int) road_data[3];
-  max_acceleration = (float) road_data[4];
-  max_legal_speed = (float) road_data[5];
+  max_acceleration = (double) road_data[4];
+  max_legal_speed = (double) road_data[5];
 }
 
 int Vehicle::get_lane() {
@@ -307,20 +306,20 @@ int Vehicle::get_lane() {
   return Vehicle::d_to_lane(this->d);
 }
 
-float Vehicle::lane_to_d(float lane) {
+double Vehicle::lane_to_d(double lane) {
   // Lanes are 4m wide, so first multiply by 4 and then add 2 to center it.
   // assumes 0 is the left most lane.
   return lane*4 + 2;
 }
 
-int Vehicle::d_to_lane(float d) {
+int Vehicle::d_to_lane(double d) {
   // Lanes are 4m wide, so first multiply by 4 and then add 2 to center it.
   // assumes 0 is the left most lane.
   return (int) round(d / 4 - 0.5);
 }
 
 bool Vehicle::in_my_lane(Vehicle &other) {
-  //const float lane_buffer = 0.5;
+  //const double lane_buffer = 0.5;
   int other_lane = other.get_lane();
   if (other_lane < 0) return false;
   int my_lane = this->get_lane();
@@ -330,23 +329,23 @@ bool Vehicle::in_my_lane(Vehicle &other) {
   ;
 }
 
-float Vehicle::distance_from_me(Vehicle &other, float time_delta) {
-  float other_car_s = other.s;
+double Vehicle::distance_from_me(Vehicle &other, double time_delta) {
+  double other_car_s = other.s;
   other_car_s += (time_delta *  other.v); //if using time_delta, project car's s value out in time.
   return (other_car_s - this->s);
 }
 
 
-float Vehicle::accelerate(float factor) {
-  float new_a = this->a + factor;
+double Vehicle::accelerate(double factor) {
+  double new_a = this->a + factor;
   if (abs(new_a) <= this->max_acceleration) {
     this->a = new_a;
   }
   return new_a;
 }
 
-float Vehicle::decelerate(float factor) {
-  float new_a = this->a - factor;
+double Vehicle::decelerate(double factor) {
+  double new_a = this->a - factor;
   if (abs(new_a) <= this->max_acceleration) {
     this->a = new_a;
   }
